Compress a bit more
