+++ dialogue-clients/build.gradle
@@ -5,6 +5,7 @@ dependencies {
     api 'com.palantir.refreshable:refreshable'
     api 'com.palantir.conjure.java.api:service-config'
     api project(':dialogue-target') // users might want to provide a ConjureRuntime
+    api 'com.palantir.conjure.java.runtime:client-config' // NodeSelectionStrategy etc
 
     implementation project(':dialogue-apache-hc4-client')
     implementation project(':dialogue-core')
@@ -13,7 +14,6 @@ dependencies {
     implementation 'com.google.code.findbugs:jsr305'
     implementation 'com.google.errorprone:error_prone_annotations'
     implementation 'com.google.guava:guava'
-    implementation 'com.palantir.conjure.java.runtime:client-config'
     implementation 'com.palantir.conjure.java.runtime:keystores'
     implementation 'com.palantir.safe-logging:preconditions'
     implementation 'com.palantir.safe-logging:safe-logging'
