--- dialogue-clients/src/main/java/com/palantir/dialogue/clients/AlwaysThrowing.java
@@ -1,38 +1,0 @@
-/*
- * (c) Copyright 2020 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.dialogue.clients;
-
-import com.google.common.util.concurrent.Futures;
-import com.google.common.util.concurrent.ListenableFuture;
-import com.palantir.dialogue.Channel;
-import com.palantir.dialogue.Endpoint;
-import com.palantir.dialogue.Request;
-import com.palantir.dialogue.Response;
-import java.util.function.Supplier;
-
-final class AlwaysThrowing implements Channel {
-    private final Supplier<? extends Throwable> exceptionSupplier;
-
-    AlwaysThrowing(Supplier<? extends Throwable> exceptionSupplier) {
-        this.exceptionSupplier = exceptionSupplier;
-    }
-
-    @Override
-    public ListenableFuture<Response> execute(Endpoint _endpoint, Request _request) {
-        return Futures.immediateFailedFuture(exceptionSupplier.get());
-    }
-}
